// src/protocols/filter/comparator-operator-enum.ts
export enum ComparatorOperatorEnum {
  EQ = 'EQ',
  EQ_ID = 'EQ_ID',
  NEQ = 'NEQ',
  NEQ_ID = 'NEQ_ID',
  GT = 'GT',
  GTE = 'GTE',
  LT = 'LT',
  LTE = 'LTE',
  IN = 'IN',
  IN_ID = 'IN_ID',
  NIN = 'NIN',
  EXISTS = 'EXISTS',
  REGEX = 'REGEX',
  BETWEEN = 'BETWEEN',
  OP_AND = 'OP_AND',
  OP_NOT = 'OP_NOT',
  OP_NOR = 'OP_NOR',
  OP_OR = 'OP_OR'
}